+++ src/main/java/com/example/samplesalad/controller/SignupController.java
@@ -1,8 +1,8 @@
 package com.example.samplesalad.controller;
 
-import com.example.samplesalad.model.UserDAO;
-import com.example.samplesalad.model.UserService;
-import com.example.samplesalad.model.UserValidator;
+import com.example.samplesalad.model.DAO.UserDAO;
+import com.example.samplesalad.model.user.UserService;
+import com.example.samplesalad.model.user.UserValidator;
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
 import javafx.scene.control.PasswordField;
