+++ src/main/java/com/example/samplesalad/model/DAO/ProjectDAO.java
@@ -1,4 +1,7 @@
-package com.example.samplesalad.model;
+package com.example.samplesalad.model.DAO;
+
+import com.example.samplesalad.model.DatabaseConnection;
+import com.example.samplesalad.model.Project;
 
 import java.sql.*;
 import java.util.ArrayList;
