+++ src/main/java/com/example/samplesalad/model/DAO/ProjectSequenceDAO.java
@@ -1,4 +1,7 @@
-package com.example.samplesalad.model;
+package com.example.samplesalad.model.DAO;
+
+import com.example.samplesalad.model.Project;
+import com.example.samplesalad.model.Sequencer;
 
 import java.sql.*;
 import java.util.ArrayList;
