+++ src/test/java/UserTest.java
@@ -1,4 +1,4 @@
-import com.example.samplesalad.model.User;
+import com.example.samplesalad.model.user.User;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
